Added function that finds pixels' shade regions

+++ src/postprocessing.rs
@@ -1,6 +1,6 @@
 use image::{ImageBuffer, LumaA};
 
 /// Saves an image.
-pub fn save_img(img: ImageBuffer<LumaA<u16>, Vec<u16>>) -> () {
+pub fn save_img(img: ImageBuffer<LumaA<u8>, Vec<u8>>) -> () {
     img.save("./out/out_img.png").unwrap();
 }
